requirements.txt
django==1.11
django-mptt==0.8.7